asr_challenge/requirements.txt
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==7.4.0
python-magic==0.4.27
yara-python==4.3.1
